requirements.txt
# Runtime dependencies
ollama==0.6.1
instructor==1.14.4
openai==2.15.0
nicegui==3.5.0
pydantic==2.12.5
networkx==3.6.1
PyYAML==6.0.2
ebooklib==0.20
reportlab==4.4.9
Pillow==12.1.0
python-docx==1.2.0

# Testing
pytest==9.0.2
pytest-cov==7.0.0
pytest-asyncio==1.3.0
pytest-xdist==3.8.0
playwright==1.57.0
pytest-playwright==0.7.2

# Code quality
ruff==0.14.13
mypy==1.19.1
pre-commit==4.5.1

# Type stubs
types-networkx==3.6.1.20251220
types-Pillow==10.2.0.20240822
types-reportlab==4.4.9.20260117
